ref: Moved model to another package
