+++ scripts/scheme.py
@@ -1,4 +1,4 @@
-from template import TemplateScheme
+from model.template import TemplateScheme
 
 if __name__ == "__main__":
     print(TemplateScheme.schema_json())
